Refactor Pagination component and improve profile management
- Moved pagination logic to a new utility function for better separation of concerns and code reusability.
- Updated Pagination component to utilize the new getPageNumbers function.
- Removed email state and related UI elements from useProfile and Profile components to streamline user data management.
- Adjusted styles in pagination for improved visual consistency.
