+++ app/utils/pagination.ts
@@ -1,0 +1,32 @@
+export default function getPageNumbers(currentPage: number, totalPages: number): number[] {
+  const pageNumbers: number[] = [];
+
+  if (totalPages <= 5) {
+    for (let i = 1; i <= totalPages; i++) {
+      pageNumbers.push(i);
+    }
+  } else {
+    pageNumbers.push(1);
+
+    if (currentPage > 3) {
+      pageNumbers.push(-1);
+    }
+
+    const start = Math.max(2, currentPage - 1);
+    const end = Math.min(totalPages - 1, currentPage + 1);
+
+    for (let i = start; i <= end; i++) {
+      pageNumbers.push(i);
+    }
+
+    if (currentPage < totalPages - 2) {
+      pageNumbers.push(-1);
+    }
+
+    if (totalPages > 1) {
+      pageNumbers.push(totalPages);
+    }
+  }
+
+  return pageNumbers;
+}
